Rename component to AutomatonAdditionalProperties
