+++ src/components/3-organisms/automaton/AutomatonAdditionalPropertiesStyle.js
